+++ src/aggregate/decl_pygments.py
@@ -134,6 +134,7 @@ class AggLexer(RegexLexer):
                     'netceded', 'grossceded', 'grossnet',
                     'approximate', 'approx', 'ssev',
                     'claims', 'ceded', 'claim', 'loss', 'payoff', 'dist',
+                    'expense', 'expenses', 'cede', 'deposit', 'rol',
                     'port', 'rate', 'net', 'sev', 'agg', 'xps', 'wts',
                     'inf', 'and', 'exp', 'at', 'cv', 'lr', 'xs',
                     'of', 'to', 'po', 'so', 'zm', 'zt', ),